Update README.md
Use the mouse to direct the main character around the screen to eat the bee's. Try to avoid the lightening bug as he will electrocute your player. Collect hearts as they come to gain lives. When you have eaten and rid the world of all the butter beez you win!